requirements.txt
fastapi
uvicorn
jinja2
google-generativeai
python-decouple
sqlalchemy[asyncio]
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
python-multipart